Enable users to subscribe to a limited set of other users presences.
No data is transmitted to clients that they are not meant to receive.

Presence state is not duplicated server side.

This is done through separating the tracked channel/process from the
presence topic. Any interested party (such as a friends user socket) may
subscribe to the relevant presence topics and transmit their data to
the user in an appropriate fashion.