Add slow fetch/2 callback for Phoenix.Presence.
This commit delays the fetch call with a full second for each presence.
When using MyPresence.list, it will be very slow after just a few
presences exist.

This displays the possibility of a race condition in presence. If
a presence_diff with leaves are pushed to the client before a presence_state
event based on data where the presence was present, those leaves will be
forgotten and kept in the presence state forever (or until the next
presence_state event, which normally only occurs on reconnects).